+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/data/response/shippingeditor/GetShipperTickerResponse.kt
@@ -107,14 +107,14 @@ data class Warehouses(
 
 data class ShopId(
         @SerializedName("int64")
-        var int64: Int = 0,
+        var int64: Long = 0,
         @SerializedName("valid")
         var valid: Boolean = false
 )
 
 data class PartnerId(
         @SerializedName("int64")
-        var int64: Int = 0,
+        var int64: Long = 0,
         @SerializedName("valid")
         var valid: Boolean = false
 )
